refactor(views): split ReflowableBookView into ReflowableRenderer + ColumnRenderer + ScrollRenderer
Retires the single ReflowableBookView class (whose setMode toggle
mutated state in place) for a renderer-per-mode architecture:

- ReflowableRenderer (abstract) — shared state and helpers (iframe,
  host, attributes, sizing, scrollingElement, getProperty, stop,
  goToCssSelector, goToFragment, getScreenHeight)
- ColumnRenderer — paginated columns. engage() applies multi-column
  CSS + sizing. Carries forward the fix from the prior commit (no
  padOddColumns call during pre-settled layout).
- ScrollRenderer — vertical-scroll mode. engage() applies scroll CSS,
  toggles iframe \`scrolling\` attribute based on scrollContainerMode,
  delegates iframe sizing to setIframeHeight.

Wiring:
- UserSettings: ColumnRenderer/ScrollRenderer/FixedRenderer imported.
  Constructor instantiates ColumnRenderer for default reflowable.
  Replaces three \`view.setMode()\` call sites with a new
  \`swapRenderer(scroll)\` helper that constructs a fresh renderer of
  the correct class on mode change, transfers iframe/host/attributes/
  sizing state, and fires viewChangeCallback so listeners re-sync.
- EpubNavigator.updateRenderer: re-syncs \`this.view\` from
  \`settings.view\` at top so the navigator follows swap-induced
  instance changes.
- Renderer interface: setMode? removed (no longer used externally;
  swapRenderer replaces it).

ReflowableBookView.ts deleted.
