--- src/views/ReflowableBookView.ts
@@ -1,692 +1,0 @@
-/*
- * Copyright 2018-2020 DITA (AM Consulting LLC)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Developed on behalf of: DITA
- * Licensed to: Bokbasen AS and CAST under one or more contributor license agreements.
- */
-
-import { UserProperty } from "../model/user-settings/UserProperties";
-import * as HTMLUtilities from "../utils/HTMLUtilities";
-import * as BrowserUtilities from "../utils/BrowserUtilities";
-import Store from "../store/Store";
-import Renderer, { RendererHost } from "./Renderer";
-import { IFrameAttributes } from "../navigator/types";
-import debounce from "debounce";
-
-export default class ReflowableBookView implements Renderer {
-  layout = "reflowable";
-
-  private readonly USERSETTINGS = "userSetting";
-  private readonly store: Store;
-  private scrollMode: boolean;
-  host: RendererHost;
-  constructor(store: Store) {
-    this.store = store;
-
-    if (this.scrollMode) {
-      this.name = "readium-scroll-on";
-      this.label = "Scrolling";
-    } else {
-      this.name = "readium-scroll-off";
-      this.label = "Paginated";
-    }
-  }
-
-  setMode(scroll: boolean) {
-    this.scrollMode = scroll;
-    if (scroll) {
-      this.name = "readium-scroll-on";
-      this.label = "Scrolling";
-      let doc = this.iframe.contentDocument;
-      if (doc) {
-        const head = HTMLUtilities.findIframeElement(
-          doc,
-          "head"
-        ) as HTMLHeadElement;
-        if (head) {
-          const viewport = HTMLUtilities.findElement(
-            head,
-            "meta[name=viewport]"
-          );
-          if (viewport) {
-            viewport.remove();
-          }
-        }
-        const html = HTMLUtilities.findIframeElement(
-          doc,
-          "html"
-        ) as HTMLHtmlElement;
-        if (html) {
-          html.style.setProperty("--USER__scroll", "readium-scroll-on");
-        }
-        // Remove column spacer — it creates whitespace in scroll mode
-        const spacer = doc.getElementById("r2d2bc-column-spacer");
-        if (spacer) spacer.remove();
-      }
-      // In iframe-scroll mode, the iframe stays at viewport height and its
-      // own document scrolls. Compute the target height the same way the
-      // paginated path does so the iframe fits the wrapper minus chrome.
-      if (this.scrollContainerMode === "iframe") {
-        this.height = BrowserUtilities.computeIframeContentHeight(
-          this.iframe,
-          this.attributes
-        );
-      }
-      this.setSize();
-      this.setIframeHeight(this.iframe);
-    } else {
-      this.height = BrowserUtilities.computeIframeContentHeight(
-        this.iframe,
-        this.attributes
-      );
-      this.name = "readium-scroll-off";
-      this.label = "Paginated";
-      // any is necessary because CSSStyleDeclaration type does not include
-      // all the vendor-prefixed attributes.
-
-      this.checkForFixedScrollWidth();
-
-      let doc = this.iframe.contentDocument;
-      if (doc) {
-        const html = HTMLUtilities.findIframeElement(
-          doc,
-          "html"
-        ) as HTMLHtmlElement;
-        if (html) {
-          html.style.setProperty("--USER__scroll", "readium-scroll-off");
-        }
-      }
-      this.setSize();
-      // padOddColumns intentionally NOT called here. Pre-settled layout at
-      // this point (mid-applyProperties, before fonts.ready / images, before
-      // multi-column flow has stabilized) produces non-deterministic
-      // remainder math — sometimes injects a spacer that isn't needed. The
-      // call sites in handleIFrameLoad / navigate / hideLoadingMessage all
-      // fire after layout has settled and handle this correctly.
-    }
-    if (this.host.isContentProtectionEnabled()) {
-      this.host.recalculateContentProtection();
-    }
-  }
-
-  name: string;
-  label: string;
-  iframe: HTMLIFrameElement;
-  sideMargin: number = 20;
-  height: number = 0;
-  attributes: IFrameAttributes = {};
-
-  start(): void {
-    if (this.scrollMode) {
-      let doc = this.iframe.contentDocument;
-      if (doc) {
-        const head = HTMLUtilities.findIframeElement(
-          doc,
-          "head"
-        ) as HTMLHeadElement;
-
-        if (head) {
-          const viewport = HTMLUtilities.findElement(
-            head,
-            "meta[name=viewport]"
-          );
-          if (viewport) {
-            viewport.remove();
-          }
-        }
-      }
-      this.setSize();
-      this.setIframeHeight(this.iframe);
-    } else {
-      this.iframe.height = "0";
-      this.iframe.width = "0";
-
-      // any is necessary because CSSStyleDeclaration type does not include
-      // all the vendor-prefixed attributes.
-      this.setSize();
-      const viewportElement = document.createElement("meta");
-      viewportElement.name = "viewport";
-      viewportElement.content =
-        "width=device-width, initial-scale=1, maximum-scale=1";
-
-      this.checkForFixedScrollWidth();
-    }
-  }
-
-  stop(): void {
-    this.iframe.height = "0";
-    this.iframe.width = "0";
-    let doc = this.iframe.contentDocument;
-    if (doc) {
-      const body = HTMLUtilities.findIframeElement(
-        this.iframe.contentDocument,
-        "body"
-      ) as HTMLBodyElement;
-      const images = Array.prototype.slice.call(body.querySelectorAll("img"));
-      for (const image of images) {
-        image.style.maxWidth = "";
-      }
-    }
-  }
-
-  getCurrentPosition(): number {
-    if (this.scrollMode) {
-      const extent = this.getScrollExtent();
-      return extent > 0 ? this.getScrollOffset() / extent : 0;
-    } else {
-      const width = this.getColumnWidth();
-      const leftWidth = this.getLeftColumnsWidth();
-      const rightWidth = this.getRightColumnsWidth();
-      const totalWidth = leftWidth + width + rightWidth;
-      return leftWidth / totalWidth;
-    }
-  }
-
-  goToProgression(position: number): void {
-    if (this.scrollMode) {
-      this.setScrollOffset(this.getScrollExtent() * position);
-      return;
-    }
-    // If the window has changed size since the columns were set up,
-    // we need to reset position, so we can determine the new total width.
-
-    const width = this.getColumnWidth();
-    const leftWidth = this.getLeftColumnsWidth();
-    const rightWidth = this.getRightColumnsWidth();
-    const totalWidth = leftWidth + width + rightWidth;
-
-    const newLeftWidth = position * totalWidth;
-
-    // Round the new left width, so it's a multiple of the column width.
-
-    let roundedLeftWidth = Math.round(newLeftWidth / width) * width;
-    if (roundedLeftWidth >= totalWidth) {
-      // We've gone too far and all the columns are off to the left.
-      // Move one column back into the viewport.
-      roundedLeftWidth = roundedLeftWidth - width;
-    }
-    this.setLeftColumnsWidth(roundedLeftWidth);
-  }
-
-  goToCssSelector(cssSelector: string, relative?: boolean): void {
-    let doc = this.iframe.contentDocument;
-    if (doc) {
-      let element = doc.querySelector(cssSelector) as HTMLElement;
-      this.goToElement(element, relative);
-    }
-  }
-
-  goToFragment(fragment: string, relative?: boolean): void {
-    let doc = this.iframe.contentDocument;
-    if (doc) {
-      const element = doc.getElementById(fragment) as HTMLElement;
-      this.goToElement(element, relative);
-    }
-  }
-
-  snap(element: HTMLElement | null, _relative?: boolean): void {
-    if (element) {
-      if (this.scrollMode) {
-        //     // Put the element as close to the top as possible.
-        //     document.scrollingElement.scrollTop = element.offsetTop;
-      } else {
-        // Get the element's position in the iframe, and
-        // round that to figure out the column it's in.
-        // There is a bug in Safari when using getBoundingClientRect
-        // on an element that spans multiple columns. Temporarily
-        // set the element's height to fit it on one column, so we
-        // can determine the first column position.
-        const originalHeight = element.style.height;
-        element.style.height = "0";
-
-        const width = this.getColumnWidth();
-        const left =
-          this.getLeftColumnsWidth() + element.getBoundingClientRect().left;
-        let roundedLeftWidth = Math.floor(left / width) * width;
-
-        // Restore element's original height.
-        element.style.height = originalHeight;
-        this.setLeftColumnsWidth(roundedLeftWidth);
-
-        if (this.host.isContentProtectionEnabled()) {
-          this.host.recalculateContentProtection(0);
-        }
-      }
-    }
-  }
-
-  goToElement(element: HTMLElement | null, relative?: boolean): void {
-    if (this.scrollMode) {
-      if (element) {
-        element.scrollIntoView({ block: "center" });
-      }
-    } else {
-      if (element) {
-        // Get the element's position in the iframe, and
-        // round that to figure out the column it's in.
-
-        // There is a bug in Safari when using getBoundingClientRect
-        // on an element that spans multiple columns. Temporarily
-        // set the element's height to fit it on one column, so we
-        // can determine the first column position.
-        const originalHeight = element.style.height;
-        element.style.height = "0";
-
-        const left = element.getBoundingClientRect().left;
-        const width = this.getColumnWidth();
-        const diff = this.scrollingElement.scrollLeft - width;
-        // let roundedLeftWidth = Math.ceil(left / width) * width;
-        let roundedLeftWidth = Math.ceil(left / width) * width + diff;
-        if (relative) {
-          const origin = this.getLeftColumnsWidth();
-          roundedLeftWidth = Math.ceil(left / width) * width + origin;
-        }
-
-        // Restore element's original height.
-        element.style.height = originalHeight;
-        this.setLeftColumnsWidth(roundedLeftWidth);
-        if (this.host.isContentProtectionEnabled()) {
-          this.host.recalculateContentProtection(200);
-        }
-      }
-    }
-  }
-
-  // at top in scroll mode
-  atStart(): boolean {
-    if (this.scrollMode) {
-      return this.getScrollOffset() === 0;
-    } else {
-      const leftWidth = this.getLeftColumnsWidth();
-      return leftWidth <= 0;
-    }
-  }
-
-  // at bottom in scroll mode
-  atEnd(): boolean {
-    if (this.scrollMode) {
-      return (
-        Math.ceil(this.getScrollExtent() - this.getScrollOffset()) - 1 <=
-        BrowserUtilities.getHeight()
-      );
-    } else {
-      // Sub-pixel rounding: getRightColumnsWidth() can return fractional values
-      // like 0.09 due to browser layout — Math.floor normalizes to 0.
-      const rightWidth = Math.floor(this.getRightColumnsWidth());
-      if (rightWidth <= 0) return true;
-      // If a spacer was injected, don't use page count check — it can
-      // trigger too early. Only rightWidth determines the real end.
-      const doc = this.iframe.contentDocument;
-      if (doc?.getElementById("r2d2bc-column-spacer")) return rightWidth <= 5;
-      // Use Math.round to match how `displayInfo.resourceScreenIndex/Count`
-      // is computed in EpubNavigator. `Math.ceil` drifted above pageCount
-      // on sub-pixel scrollLeft variations, permanently locking atEnd to
-      // false on the last page even when the UI showed "Page N of N".
-      // Using `>=` also covers the case where currentPage computation
-      // overshoots pageCount slightly on the real last page.
-      return Math.round(this.getCurrentPage()) >= this.getPageCount();
-    }
-  }
-
-  goToPreviousPage(): void {
-    if (this.scrollMode) {
-      const leftHeight = this.getScrollOffset();
-      const height = this.getScreenHeight() - 40;
-      const offset = leftHeight - height;
-      this.setScrollOffset(offset >= 0 ? offset : 0);
-    } else {
-      const leftWidth = this.getLeftColumnsWidth();
-      const width = this.getColumnWidth();
-      const offset = leftWidth - width;
-      if (offset >= 0) {
-        this.setLeftColumnsWidth(offset);
-      } else {
-        this.setLeftColumnsWidth(0);
-      }
-      this.host.checkResourcePosition();
-    }
-    if (this.host.isContentProtectionEnabled()) {
-      this.host.recalculateContentProtection();
-    }
-  }
-
-  goToNextPage(): void {
-    if (this.scrollMode) {
-      const leftHeight = this.getScrollOffset();
-      const height = this.getScreenHeight() - 40;
-      const scrollHeight = this.getScrollExtent();
-      const offset = leftHeight + height;
-      this.setScrollOffset(offset < scrollHeight ? offset : scrollHeight);
-    } else {
-      const leftWidth = this.getLeftColumnsWidth();
-      const width = this.getColumnWidth();
-      const scrollWidth = this.scrollWidth;
-      const offset = leftWidth + width;
-      if (offset < scrollWidth) {
-        this.setLeftColumnsWidth(offset);
-      } else {
-        this.setLeftColumnsWidth(scrollWidth);
-      }
-      this.host.checkResourcePosition();
-    }
-    if (this.host.isContentProtectionEnabled()) {
-      this.host.recalculateContentProtection();
-    }
-  }
-
-  // doesn't exist in scroll mode
-  getCurrentPage(): number {
-    if (this.scrollMode) {
-      return 0;
-    } else {
-      return this.getCurrentPosition() * this.getPageCount() + 1;
-    }
-  }
-
-  // doesn't exist in scroll mode
-  getPageCount(): number {
-    if (this.scrollMode) {
-      return 0;
-    } else {
-      const width = this.getColumnWidth();
-      return this.scrollWidth / width;
-    }
-  }
-
-  isPaginated() {
-    if (this.iframe) {
-      let doc = this.iframe.contentDocument;
-      if (doc) {
-        const html = HTMLUtilities.findIframeElement(
-          doc,
-          "html"
-        ) as HTMLHtmlElement;
-        if (html) {
-          const scroll =
-            "readium-scroll-on" ===
-            html.style.getPropertyValue("--USER__scroll");
-          return !scroll;
-        }
-      }
-    }
-    return !this.scrollMode;
-  }
-
-  isScrollMode() {
-    if (this.iframe) {
-      let doc = this.iframe.contentDocument;
-      if (doc) {
-        const html = HTMLUtilities.findIframeElement(
-          doc,
-          "html"
-        ) as HTMLHtmlElement;
-        if (html) {
-          return (
-            "readium-scroll-on" ===
-            html.style.getPropertyValue("--USER__scroll")
-          );
-        }
-      }
-    }
-    return this.scrollMode;
-  }
-
-  async getProperty(name: string): Promise<UserProperty | null> {
-    let array = await this.store.get(this.USERSETTINGS);
-    if (array) {
-      let properties = JSON.parse(array) as Array<UserProperty>;
-      properties = properties.filter((el: UserProperty) => el.name === name);
-      if (properties.length === 0) {
-        return null;
-      }
-      return properties[0];
-    }
-    return null;
-  }
-
-  // scrolling functions
-
-  getScreenHeight(): number {
-    const wrapper = HTMLUtilities.findRequiredElement(
-      document,
-      "#iframe-wrapper"
-    );
-    return wrapper.clientHeight;
-  }
-
-  /**
-   * Which element provides the scrollbar in scroll mode.
-   *
-   * - `"host"` (default): `#iframe-wrapper` scrolls; iframe grows to content.
-   * - `"iframe"`: iframe's own document scrolls; iframe stays at viewport.
-   */
-  private get scrollContainerMode(): "host" | "iframe" {
-    return this.attributes?.scrollContainer === "iframe" ? "iframe" : "host";
-  }
-
-  /** Current scroll offset (px from top) of the active scroll container. */
-  private getScrollOffset(): number {
-    if (this.scrollContainerMode === "iframe") {
-      return this.scrollingElement?.scrollTop ?? 0;
-    }
-    const wrapper = HTMLUtilities.findRequiredElement(
-      document,
-      "#iframe-wrapper"
-    );
-    return wrapper.scrollTop;
-  }
-
-  /** Set scroll offset on the active scroll container. */
-  private setScrollOffset(value: number): void {
-    if (this.scrollContainerMode === "iframe") {
-      if (this.scrollingElement) this.scrollingElement.scrollTop = value;
-      return;
-    }
-    const wrapper = HTMLUtilities.findRequiredElement(
-      document,
-      "#iframe-wrapper"
-    );
-    wrapper.scrollTop = value;
-  }
-
-  /**
-   * Total scrollable height. Same value in both modes — iframe's internal
-   * document `scrollHeight`, because the iframe's document is always the
-   * source of truth for "how much content is there."
-   */
-  private getScrollExtent(): number {
-    return this.scrollingElement?.scrollHeight ?? 0;
-  }
-
-  setIframeHeight(iframe: any) {
-    // Iframe-scroll mode keeps the iframe at viewport height; sizing is
-    // owned by `setSize()` and there is nothing to debounce-grow here.
-    if (this.scrollContainerMode === "iframe") return;
-    let d = debounce((iframe: any) => {
-      if (iframe) {
-        let body = iframe.contentWindow.document.body,
-          html = iframe.contentWindow.document.documentElement;
-
-        let height = Math.max(
-          body?.scrollHeight,
-          body?.offsetHeight,
-          html?.clientHeight,
-          html?.scrollHeight,
-          html?.offsetHeight
-        );
-        if (height) {
-          const minHeight = BrowserUtilities.computeIframeContentHeight(
-            iframe,
-            this.attributes
-          );
-          iframe.height = Math.max(minHeight, height) + "px";
-        }
-      }
-    }, 200);
-    d(iframe);
-  }
-
-  // paginated functions
-
-  protected hasFixedScrollWidth: boolean = false;
-
-  protected checkForFixedScrollWidth(): void {
-    // Determine if the scroll width changes when the left position
-    // changes. This differs across browsers and sometimes across
-    // books in the same browser.
-    let doc = this.iframe.contentDocument;
-    if (doc) {
-      const body = HTMLUtilities.findIframeElement(doc, "body");
-      const originalScrollWidth = body?.scrollWidth;
-      this.hasFixedScrollWidth = body?.scrollWidth === originalScrollWidth;
-    }
-  }
-
-  setSize(): void {
-    this.iframe.width = BrowserUtilities.getWidth() + "px";
-    if (!this.scrollMode) {
-      // Paginated: iframe and its document are constrained to the same
-      // height so columns fit the viewport.
-      let doc = this.iframe.contentDocument;
-      if (doc && doc.documentElement) {
-        doc.documentElement.style.height = this.height + "px";
-      }
-      this.iframe.height = this.height + "px";
-    } else if (this.scrollContainerMode === "iframe") {
-      // Iframe-scroll: iframe stays at viewport height; its document is
-      // intentionally unconstrained so it overflows and scrolls inside
-      // the iframe.
-      this.iframe.height = this.height + "px";
-    } else {
-      // Host-scroll: iframe grows to content; #iframe-wrapper scrolls.
-      let html = this.iframe.contentWindow?.document?.documentElement;
-      this.iframe.height = html?.offsetHeight + "px";
-    }
-  }
-
-  // TODO: if we changed the following functions to handle screen size (height and width) we can use it for both paginated and reflowable.
-  // For example: getColumnWidth would be renamed to something that gives us either the offsetWidth when in paginated mode and the offsetHeight when in scroll mode. etc.
-  // Since theses are the only 4 functions right now that are specifically used for paginated mode but gives us nothing for scroll mode, it would make sense to make them more universal
-
-  /** Returns the total width of the columns that are currently
-     positioned to the left of the iframe viewport. */
-  private getLeftColumnsWidth(): number {
-    return Math.round(this.scrollingElement.scrollLeft);
-  }
-
-  /** Returns the total width of the columns that are currently
-     positioned to the right of the iframe viewport. */
-  private getRightColumnsWidth(): number {
-    const width = this.getColumnWidth();
-    let rightWidth = this.scrollWidth - width;
-    if (this.hasFixedScrollWidth) {
-      // In some browsers (IE and Firefox with certain books),
-      // scrollWidth doesn't change when some columns
-      // are off to the left, so we need to subtract them.
-      const leftWidth = this.getLeftColumnsWidth();
-      rightWidth = Math.max(0, rightWidth - leftWidth);
-    }
-    return rightWidth;
-  }
-
-  /** Returns the width of one column. */
-  private getColumnWidth(): number {
-    return this.scrollingElement.clientWidth;
-  }
-
-  /** Shifts the columns so that the specified width is positioned
-     to the left of the iframe viewport. */
-  private setLeftColumnsWidth(width: number) {
-    this.scrollingElement.scrollLeft = width;
-  }
-
-  get scrollingElement() {
-    if (this.iframe.contentDocument?.scrollingElement) {
-      return this.iframe.contentDocument?.scrollingElement;
-    } else if (this.iframe.contentDocument?.body) {
-      return this.iframe.contentDocument?.body;
-    } else {
-      // Note: this should never happen, but prevents any exceptions.
-      // exception here could happen id the iframe s not loaded yet properly
-      // but the scrolling element or body is accessed
-      // for example, very fast next/previous resource changes
-      return document.createElement("body");
-    }
-  }
-
-  get scrollWidth() {
-    const scrollWidth = this.scrollingElement?.scrollWidth;
-    const width = this.getColumnWidth();
-    // Subtract 1px before ceil to absorb sub-pixel rounding that
-    // creates phantom pages (html/body scrollWidth off by 1-2px)
-    const pages = Math.ceil((scrollWidth - 1) / width);
-    return Math.max(1, pages) * width;
-  }
-
-  /**
-   * In multi-column paginated mode, if the content produces an odd number
-   * of CSS columns the last page will only show one column and navigation
-   * may skip it. This injects an invisible spacer to force an even column
-   * count so every page is full.
-   */
-  padOddColumns(): boolean {
-    if (this.scrollMode) return false;
-
-    const doc = this.iframe.contentDocument;
-    if (!doc) return false;
-
-    // If a spacer already exists, don't re-evaluate
-    const existing = doc.getElementById("r2d2bc-column-spacer");
-    if (existing) return false;
-
-    const body = doc.body;
-    if (!body) return false;
-
-    const rawScrollWidth = body.scrollWidth;
-    const viewportWidth = this.getColumnWidth();
-    const remainder = rawScrollWidth % viewportWidth;
-
-    // Guard against floated descendants (e.g. `float: right` images) which
-    // can inflate body.scrollWidth past the real content extent. Measure
-    // where the last real child actually ends; if that is clearly short of
-    // scrollWidth, the remainder we see is a float artifact, not a genuine
-    // odd column. Skip the spacer in that case.
-    const bodyLeft = body.getBoundingClientRect().left;
-    const last = body.lastElementChild as HTMLElement | null;
-    if (last) {
-      const realContentRight = last.getBoundingClientRect().right - bodyLeft;
-      // If real content ends more than half a viewport short of the declared
-      // scrollWidth, the tail is inflated by a float or similar.
-      if (rawScrollWidth - realContentRight > viewportWidth * 0.5) {
-        return false;
-      }
-    }
-
-    // Only inject spacer when remainder is at least 25% of viewport —
-    // an actual odd column is ~50% of viewport
-    if (remainder > viewportWidth * 0.25) {
-      const spacer = doc.createElement("div");
-      spacer.id = "r2d2bc-column-spacer";
-      spacer.style.breakBefore = "column";
-      spacer.style.height = "100%";
-      spacer.style.visibility = "hidden";
-      body.appendChild(spacer);
-      // Force synchronous reflow
-      void body.scrollWidth;
-      return true;
-    }
-    return false;
-  }
-}
